get location of command from env (#9)
